refactor: split collision files
